+++ requirements.txt
@@ -1,1 +1,4 @@
+scipy==0.18.0
+tensorflow==1.0.1
 tweepy==3.5.0
+
