+++ tests/collector_test.py
@@ -1,8 +1,14 @@
-from context import ttrends
-from ttrends.collection.DataHandler import DataHandler
-from ttrends.collection.DataCollector import DataCollector
-from ttrends.collection.StreamTransformer import *
-from ttrends.collection.AuthInfo import * # where api access information is stored
+############################################################################################
+# PROGRAMMED BY: Jean Flaherty                                                             #
+# DATE: 4-07-2017                                                                          #
+# DESCRIPTION: This script tests the data collection tools defined in the ttrends package  #
+#                                                                                          #
+############################################################################################
+
+from ttrends.collection.data_handler import DataHandler
+from ttrends.collection.data_collector import DataCollector
+from ttrends.collection.stream_transformer import *
+from ttrends.collection.auth_info import * # where api access information is stored
 
 collector = DataCollector(access_token, access_token_secret, consumer_key, consumer_secret)
 collector.authenticate()
