--- ttrends/collection/AuthInfo.py
@@ -1,4 +1,0 @@
-access_token = "3911012232-MNWJkB5E5EnN8pNrqPu8TjOIJrXHs5TmsSVI6dW"
-access_token_secret = "jjOsGntWGcxQdPZg3ZVamCdhmYjzNDOPZxMBa1zyn1Cic"
-consumer_key = "ToqLVlPLUsqrvHokKIitbi5ps"
-consumer_secret = "OSLHJZBOsu1nrCk476BAIDXiVCHEl7IbAwK8ZxjwMJWjOL4q3w"
