+++ ttrends/collection/auth_info.py
@@ -1,0 +1,10 @@
+###########################################################################
+# DESCRIPTION: This file provides the authorization information that is   #
+# used by tweepy to connect to the twitter API.                           #
+#                                                                         #
+###########################################################################
+
+access_token = "3911012232-MNWJkB5E5EnN8pNrqPu8TjOIJrXHs5TmsSVI6dW"
+access_token_secret = "jjOsGntWGcxQdPZg3ZVamCdhmYjzNDOPZxMBa1zyn1Cic"
+consumer_key = "ToqLVlPLUsqrvHokKIitbi5ps"
+consumer_secret = "OSLHJZBOsu1nrCk476BAIDXiVCHEl7IbAwK8ZxjwMJWjOL4q3w"
